// src/app/models/Progress.ts
export class Progress {
    done: number | null = null;
    revisions: number[] | null = null;
    learn: number | null = null;
    study: number | null = null;
    read: number | null = null;
    write: number | null = null;
    teach: number | null = null;

    constructor() {
        // this.learning = new Date();
    }
}